% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bernstein}
\alias{bernstein}
\title{Single cell RNA-Seq data extracted from a publication by Bernstein.}
\format{\preformatted{ num [1:5947, 1:364] -3.89 1.19 -5.82 -2.92 -1.42 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:364] "1" "1" "1" "1" ...
}}
\usage{
bernstein
}
\description{
Source: Patel, A. P. et al. Single-cell RNA-seq highlights intratumoral
heterogeneity in primary glioblastoma. Science 344, 1396–1401 (2014).
}
\details{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE57872}

Columns represent cells, rows represent genes expression values. Colnames
respresent indexes of cell clusters (known information based on the
experimental protocol). There are 364 cells and 5 clusters in this dataset.

NOTE: This dataset is already processed with gene_filter1 and
log transformation.
}
\keyword{datasets}

