% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{check_gene_filter1}
\alias{check_gene_filter1}
\title{Check the first filtering step}
\usage{
check_gene_filter1(d, d.name, min.cells, max.cells, min.reads, n.dim)
}
\arguments{
\item{d}{Expression matrix with rows as genes and columns as cells}

\item{d.name}{Name of the toy dataset. Either "quake", "sandberg", "linnarsson" or "bernstein"}

\item{min.cells}{Minimum number of cells in which a given gene is expressed}

\item{max.cells}{Maximum number of cells in which a given gene is expressed}

\item{min.reads}{Minimum number of reads per gene per cell}

\item{n.dim}{Number of dimension of the transformed distance matrix which is used
in kmeans clustering}
}
\value{
Adjusted Rand index of the clustering
}
\description{
Evaluate Adjusted Rand index based on gene_filter1() parameters
}
\examples{
check_gene_filter1(quake, "quake", 3, 3, 2, 4)
}

