% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustools.R
\name{cluster_real_data}
\alias{cluster_real_data}
\title{Cluster real data}
\usage{
cluster_real_data(d, min.cells, max.cells, min.reads, filter1, filter2, distan,
  clust, k, n.dim)
}
\arguments{
\item{d}{Single-cell RNA-Seq dataset.}

\item{min.cells}{Minimum number of cells in which a given gene is expressed.}

\item{max.cells}{Maximum number of cells in which a given gene is expressed.}

\item{min.reads}{Minimum number of reads per gene per cell.}

\item{filter1}{Defines whether or not to perform filter1 (see ?gene_filter1
for description)}

\item{filter2}{Selection method used by gene_filter2() function (either
"none", "correlation", "variance", "variance_weight", "shannon_weight")}

\item{distan}{Distance metrics for calculating a distance matrix (either
"pearson", "spearman", "euclidean", "manhattan" or "minkowski").}

\item{clust}{Distance matrix transformation method (either "pca", "spectral",
"spectral_reg" or "mds")}

\item{k}{Number of clusters}

\item{n.dim}{Number of dimension of the transformed distance matrix which is used
in kmeans clustering.}
}
\value{
kmeans results
}
\description{
Performs kmeans clustering of real data with a given combination of filtering 1,
filtering 2, distance metrics, transformation, number of clusters and number
of dimensions used in clustering.
}
\examples{
res <- cluster_real_data(quake, 3, 3, 2, "none", "spearman", "spectral", 5, 4)
}

