% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{confusion}
\alias{confusion}
\title{Confusion calculation}
\usage{
confusion(labs, labs.known)
}
\arguments{
\item{labs}{Cluster labels from test clustering}

\item{labs.known}{Cluster labels from gold standard clustering}
}
\value{
Vector containing confusions between labs and labs.known for each
cluster group
}
\description{
First test labels are matched to gold standard labels by using
minWeightBipartiteMatching function from this package. Then when labels of
labs and labs.known are of the same order, confusions between them are
calculated. Confusion metrics introduced by myself:
}
\details{
1 - number-of-matching-labels/total-number-of-labels

where number-of-matching-labels - number of labels in each cluster group of
labs that match to the labels of labs.known for the same cluster group; and
total-number-of-labels - total number of labels of labs.known from the same
cluster group.
}
\examples{
res <- confusion(labs, labs.known)
}

