% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{confusion_pipeline}
\alias{confusion_pipeline}
\title{Calclulate and plot confusions}
\usage{
confusion_pipeline(dataset, n.dim)
}
\arguments{
\item{dataset}{Name of the toy dataset.}

\item{n.dim}{Number of dimensions used for clustering of the toy dataset}
}
\value{
Two files with values of confusions and plot of the confussions
facetted by different methods
}
\description{
Only works on the output of machine_learning_pipeline function from this
package.
}
\details{
Calculate and plot confusions.
}
\examples{
confusion_pipeline("quake", 4)
}

