% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustools.R
\name{gene_filter2}
\alias{gene_filter2}
\title{Second filtering step}
\usage{
gene_filter2(d, method)
}
\arguments{
\item{d}{Expression matrix with rows as genes and columns as cells}

\item{method}{Filtering method. There are four options: "correlation",
"variance", "variance_weight", "shannon_weight" or "none". If "correlation" is used
a Pearson correlation value is used with a threshold of 0.8.}
}
\value{
Filtered expression matrix where statistically weak genes are excluded
}
\description{
Filter genes that would not contribute to clustering, because statistically they
have a very low contribution
}
\examples{
gene_filter2(quake, "correlation")
}

