% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kirschner}
\alias{kirschner}
\title{Single cell RNA-Seq data extracted from a publication by Kirschner.}
\format{\preformatted{ int [1:24174, 1:2717] 1 0 1 4 0 3 4 2 2 0 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:2717] "1" "1" "1" "1" ...
}}
\usage{
kirschner
}
\description{
Source: Klein, A. M. et al. Droplet Barcoding for Single-Cell Transcriptomics
Applied to Embryonic Stem Cells. Cell 161, 1187–1201 (2015).
}
\details{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE65525}

Columns represent cells, rows represent genes expression values. Colnames
respresent indexes of cell clusters (known information based on the
experimental protocol). There are 3005 cells and 9 clusters in this dataset.
}
\keyword{datasets}

