% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linnarsson}
\alias{linnarsson}
\title{Single cell RNA-Seq data extracted from a publication by Linnarsson.}
\format{\preformatted{ int [1:19971, 1:3005] 1 1 1 1 1 1 1 1 1 1 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:3005] "1" "1" "1" "1" ...
}}
\usage{
linnarsson
}
\description{
Source:
}
\details{
Columns represent cells, rows represent genes expression values. Colnames
respresent indexes of cell clusters (known information based on the
experimental protocol). There are 3005 cells and 9 clusters in this dataset.
}
\keyword{datasets}

