% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustools.R
\name{machine_learning_pipeline}
\alias{machine_learning_pipeline}
\title{Learning pipeline}
\usage{
machine_learning_pipeline(dataset, sel, distan, clust, n.dim)
}
\arguments{
\item{sel}{Selection method used by gene_filter2() function (either
"none", "correlation", "variance", "variance_weight", "shannon_weight")}

\item{distan}{Distance metrics for calculating a distance matrix (either
"pearson", "spearman", "euclidean", "manhattan" or "minkowski").}

\item{clust}{Distance matrix transformation method (either "pca", "spectral",
"spectral_reg" or "mds")}

\item{n.dim}{Number of dimension of the transformed distance matrix which is used
in kmeans clustering.}

\item{d}{Name of the dataset. Either "quake", "sandberg", "bernstein" or
"linnarsson"}
}
\value{
Two small files (depending on the transformation method) containing
a set of clustering evaluation indecies (*-inds.txt) together with known and
calculated clustering labels of the cells (*-labs.txt). Evaluation indecies are
the following:
\describe{
  \item{ari}{Adjusted Rand Index}
  \item{rand}{Rand Index}
  \item{jaccard}{Jaccard Index}
  \item{dunn}{Dunn Index}
  \item{davies_bouldin}{Davies Bouldin Index}
  \item{silhouette}{Silhouette Index}
}
}
\description{
Performs and evaluates kmeans clustering with a give combination of gene
filtering 2, distance metrics, transformation and number of dimensions used
in clustering.
}
\examples{
machine_learning_pipeline("quake", "none", "spearman", "spectral", 4)
}

