% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/matlab-markov-stability.R
\name{markov_stability}
\alias{markov_stability}
\title{Find structure of the data (number of clusters) based on Markov stability
method (http://github.com/michaelschaub/PartitionStability)}
\usage{
markov_stability(dataset, sel, distan, nn, matlab)
}
\arguments{
\item{dataset}{Name of the dataset. Either "quake", "sandberg", "bernstein" or
"linnarsson"}

\item{sel}{Selection method used by gene_filter2() function (either
"none", "correlation", "variance", "variance_weight", "shannon_weight")}

\item{distan}{Distance metrics for calculating a distance matrix (either
"pearson", "spearman", "euclidean", "manhattan" or "minkowski").}

\item{nn}{Number of nearest neighbours used}

\item{matlab}{Path to matlab source file (e.g. "/Applications/MATLAB_R2015a.app/bin/matlab")}
}
\value{
File dataset-sel-distan-nn-N.csv containing number of clusters at
different diffusion times. File script.m contains a sequence of commands used
in Matlab to run Markov stability algorithm.
}
\description{
WARNING: Matlab has to be installed on your computer.
}
\details{
Diffusion times are defined in the interval [10^1:10^5].
}
\examples{
markov_stability("quake", "none", "spearman", 3, "/Applications/MATLAB_R2015a.app/bin/matlab")
}

