% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{nearest_neighbour_pipeline}
\alias{nearest_neighbour_pipeline}
\title{Nearest neighbour pipeline}
\usage{
nearest_neighbour_pipeline(dataset, sel, distan, clust, n.dim, nn)
}
\arguments{
\item{dataset}{Name of the dataset. Either "quake", "sandberg", "bernstein" or
"linnarsson"}

\item{sel}{Selection method used by gene_filter2() function (either
"none", "correlation", "variance", "variance_weight", "shannon_weight")}

\item{distan}{Distance metrics for calculating a distance matrix (either
"pearson", "spearman", "euclidean", "manhattan" or "minkowski").}

\item{clust}{Distance matrix transformation method (either "pca", "spectral",
"spectral_reg" or "mds")}

\item{n.dim}{Number of dimension of the transformed distance matrix which is used
in kmeans clustering.}

\item{nn}{Number of nearest neighbours used}
}
\value{
Two small files (depending on the transformation method) containing
a set of clustering evaluation indecies (*-inds.txt) together with known and
calculated clustering labels of the cells (*-labs.txt). Evaluation indecies are
the following:
\describe{
  \item{ari}{Adjusted Rand Index}
  \item{rand}{Rand Index}
  \item{jaccard}{Jaccard Index}
  \item{dunn}{Dunn Index}
  \item{davies_bouldin}{Davies Bouldin Index}
  \item{silhouette}{Silhouette Index}
}
}
\description{
Performs and evaluates kmeans clustering with a given combination of gene
filtering 2, distance metrics, transformation, number of dimensions used in
clustering and number of nearest neighbours.
}
\examples{
nearest_neighbour_pipeline("quake", "none", "spearman", "spectral", 4, 3)
}

