% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quake}
\alias{quake}
\title{Single cell RNA-Seq data extracted from a publication by Steven Quake.}
\format{\preformatted{ num [1:23270, 1:50] 5.35 34.88 225.31 7.73 56.4 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:50] "4" "2" "4" "2" ...
}}
\usage{
quake
}
\description{
Source:
}
\details{
Columns represent cells, rows represent genes expression values. Colnames
respresent indexes of cell clusters (known information based on the
experimental protocol). There are 50 cells and 5 clusters in this dataset.
}
\keyword{datasets}

