% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sandberg}
\alias{sandberg}
\title{Single cell RNA-Seq data extracted from a publication by Sandberg.}
\format{\preformatted{ num [1:22957, 1:201] 0 151.61 0.0283 100.04 0 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:201] "1" "1" "1" "1" ...
}}
\usage{
sandberg
}
\description{
Source:
}
\details{
Columns represent cells, rows represent genes expression values. Colnames
respresent indexes of cell clusters (known information based on the
experimental protocol). There are 201 cells and 9 clusters in this dataset.
}
\keyword{datasets}

