% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{support_vector_machines}
\alias{support_vector_machines}
\title{Support vector machine (SVM) learning on toy datasets}
\usage{
support_vector_machines(dataset, teach.proportion, kern)
}
\arguments{
\item{dataset}{Name of the toy dataset}

\item{teach.proportion}{Proportion of cells used for teaching, the rest of the
cells are used for studying}

\item{kern}{The kernel used in training and predicting: linear, polynomal, radial,
sigmoid. See ?svm from "clue" package for more information.}
}
\value{
#' \describe{
  \item{pred}{Predicted cell labels for cells used in study}
  \item{ari}{Adjusted Rand Index - comparison of pred wiht gold standard}
  \item{model}{model parameters used in SVM}
  \item{training}{Training dataset}
}
}
\description{
Perform SVM learning on either quake, sandberg or linnarsson data sets.
}
\examples{
res <- support_vector_machines("sandberg", 0.7)
}

