% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zhong}
\alias{zhong}
\title{Single cell RNA-Seq data extracted from a publication by Zhong.}
\format{\preformatted{ num [1:25737, 1:49] 25.8 93.4 0 38 0 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:49] "1" "1" "1" "1" ...
}}
\usage{
zhong
}
\description{
Source: Biase, F. H., Cao, X. & Zhong, S. Cell fate inclination within 2-cell
and 4-cell mouse embryos revealed by single-cell RNA sequencing. Genome Res.
24, 1787–1796 (2014).
}
\details{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE57249}

Columns represent cells, rows represent genes expression values. Colnames
respresent indexes of cell clusters (known information based on the
experimental protocol). There are 49 cells and 3 clusters in this dataset.
}
\keyword{datasets}

