% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linnarsson}
\alias{linnarsson}
\title{Single cell RNA-Seq data extracted from a publication by Linnarsson.}
\format{\preformatted{ int [1:19972, 1:3005] 0 3 3 0 1 0 0 11 1 0 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:19972] "Tspan12" "Tshz1" "Fnbp1l" "Adamts15" ...
  ..$ : chr [1:3005] "1" "1" "1" "1" ...
}}
\usage{
linnarsson
}
\description{
Source: Zeisel, A. et al. Brain structure. Cell types in the mouse cortex and
hippocampus revealed by single-cell RNA-seq. Science 347, 1138–1142 (2015).
}
\details{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE60361}

Columns represent cells, rows represent genes expression values. Colnames
respresent indexes of cell clusters (known information based on the
experimental protocol). There are 3005 cells and 9 clusters in this dataset.
}
\keyword{datasets}

